+++ src/mlpack/bindings/cli/in_place_copy.hpp
@@ -31,17 +31,17 @@ template<typename T>
 void InPlaceCopyInternal(
     util::ParamData& /* d */,
     util::ParamData& /* input */,
-    const typename boost::disable_if<arma::is_arma_type<T>>::type* = 0,
-    const typename boost::disable_if<data::HasSerialize<T>>::type* = 0,
-    const typename boost::disable_if<std::is_same<T,
-        std::tuple<mlpack::data::DatasetInfo, arma::mat>>>::type* = 0)
+    const typename std::enable_if<!arma::is_arma_type<T>::value>::type* = 0,
+    const typename std::enable_if<!data::HasSerialize<T>::value>::type* = 0,
+    const typename std::enable_if<!std::is_same<T,
+        std::tuple<mlpack::data::DatasetInfo, arma::mat>>::value>::type* = 0)
 {
   // Nothing to do.
 }
 
 /**
  * Modify the filename for any type that needs to be loaded from disk to match
- * the filename of the input parameter.
+ * the filename of the input parameter, for a matrix/DatasetInfo parameter.
  *
  * @param d ParamData object we want to make into an in-place copy.
  * @param input ParamData object whose filename we should copy.
@@ -53,11 +53,34 @@ void InPlaceCopyInternal(
     const typename std::enable_if<
         arma::is_arma_type<T>::value ||
         std::is_same<T,
-                     std::tuple<mlpack::data::DatasetInfo, arma::mat>>::value ||
-        data::HasSerialize<T>::value>::type* = 0)
+                     std::tuple<mlpack::data::DatasetInfo, arma::mat>>::value
+                 >::type* = 0)
 {
   // Make the output filename the same as the input filename.
   typedef std::tuple<T, typename ParameterType<T>::type> TupleType;
+  TupleType& tuple = *boost::any_cast<TupleType>(&d.value);
+  std::string& value = std::get<0>(std::get<1>(tuple));
+
+  const TupleType& inputTuple = *boost::any_cast<TupleType>(&input.value);
+  value = std::get<0>(std::get<1>(inputTuple));
+}
+
+/**
+ * Modify the filename for any type that needs to be loaded from disk to match
+ * the filename of the input parameter. For serializable objects.
+ *
+ * @param d ParamData object we want to make into an in-place copy.
+ * @param input ParamData object whose filename we should copy.
+ */
+template<typename T>
+void InPlaceCopyInternal(
+    util::ParamData& d,
+    util::ParamData& input,
+    const typename std::enable_if<
+        data::HasSerialize<T>::value>::type* = 0)
+{
+  // Make the output filename the same as the input filename.
+  typedef std::tuple<T*, typename ParameterType<T>::type> TupleType;
   TupleType& tuple = *boost::any_cast<TupleType>(&d.value);
   std::string& value = std::get<1>(tuple);