+++ src/mlpack/methods/ann/layer/lp_pooling.hpp
@@ -161,12 +161,22 @@ class LpPooling
   template<typename eT>
   void Pooling(const arma::Mat<eT>& input, arma::Mat<eT>& output)
   {
+    arma::Mat<eT> inputPre = input;
+    inputPre = arma::pow(inputPre, normType);
+
+    for (size_t i = 1; i < input.n_cols; ++i)
+      inputPre.col(i) += inputPre.col(i - 1);
+
+    for (size_t i = 1; i < input.n_rows; ++i)
+      inputPre.row(i) += inputPre.row(i - 1);
+
     for (size_t j = 0, colidx = 0; j < output.n_cols;
          ++j, colidx += strideHeight)
     {
       for (size_t i = 0, rowidx = 0; i < output.n_rows;
            ++i, rowidx += strideWidth)
       {
+        double val = 0.0;
         size_t rowEnd = rowidx + kernelWidth - 1;
         size_t colEnd = colidx + kernelHeight - 1;
 
@@ -175,14 +185,22 @@ class LpPooling
         if (colEnd > input.n_cols - 1)
           colEnd = input.n_cols - 1;
 
-        arma::mat subInput = input(
-            arma::span(rowidx, rowEnd),
-            arma::span(colidx, colEnd));
+        val += inputPre(rowEnd, colEnd);
+        if (rowidx >= 1)
+        {
+          if (colidx >= 1)
+            val += inputPre(rowidx - 1, colidx - 1);
+          val -= inputPre(rowidx - 1, colEnd);
+        }
 
-        output(i, j) = pow(arma::accu(arma::pow(subInput,
-            normType)), 1.0 / normType);
+        if (colidx >= 1)
+          val -= inputPre(rowEnd, colidx - 1);
+
+        output(i, j) = val;
       }
     }
+
+    output = arma::pow(output, 1.0 / normType);
   }
 
   /**
@@ -196,11 +214,12 @@ class LpPooling
                  const arma::Mat<eT>& error,
                  arma::Mat<eT>& output)
   {
-
     arma::Mat<eT> unpooledError;
-    for (size_t j = 0, colidx = 0; j < input.n_cols; j += strideHeight, colidx++)
+    for (size_t j = 0, colidx = 0; j < input.n_cols; j += strideHeight,
+         colidx++)
     {
-      for (size_t i = 0, rowidx = 0; i < input.n_rows; i += strideWidth, rowidx++)
+      for (size_t i = 0, rowidx = 0; i < input.n_rows; i += strideWidth,
+           rowidx++)
       {
         size_t rowEnd = i + kernelWidth - 1;
         size_t colEnd = j + kernelHeight - 1;
@@ -219,7 +238,8 @@ class LpPooling
           colEnd = input.n_cols - 1;
         }
 
-        arma::mat InputArea = input(arma::span(i, rowEnd), arma::span(j, colEnd));
+        arma::mat InputArea = input(arma::span(i, rowEnd),
+            arma::span(j, colEnd));
 
         size_t sum = pow(arma::accu(arma::pow(InputArea, normType)),
             (normType - 1) / normType);