+++ src/mlpack/methods/mean_shift/mean_shift.hpp
@@ -17,7 +17,6 @@
 #include <mlpack/core/kernels/gaussian_kernel.hpp>
 #include <mlpack/core/kernels/kernel_traits.hpp>
 #include <mlpack/core/metrics/lmetric.hpp>
-#include <boost/utility.hpp>
 
 namespace mlpack {
 namespace meanshift /** Mean shift clustering. */ {