+++ src/mlpack/tests/hoeffding_tree_test.cpp
@@ -1044,7 +1044,7 @@ TEST_CASE("BatchTrainingTest", "[HoeffdingTreeTest]")
   // able to have enough samples to build to the same leaves.
   HoeffdingTree<> batchTree(trainingData, info, trainingLabels, 5, true,
       0.99999999);
-  HoeffdingTree<> streamTree(trainingLabels, info, trainingLabels, 5, false,
+  HoeffdingTree<> streamTree(trainingData, info, trainingLabels, 5, false,
       0.99999999);
 
   // Ensure that the performance of the batch tree is better.
@@ -1474,4 +1474,55 @@ TEST_CASE("HoeffdingTreeModelSerializationTest", "[HoeffdingTreeTest]")
       REQUIRE(probabilities[i] == Approx(probabilitiesBinary[i]).epsilon(1e-7));
     }
   }
+}
+
+TEST_CASE("HoeffdingTreeEmptyConstructorTrainTest", "[HoeffdingTreeTest]")
+{
+  // Generate data.
+  arma::mat data(5, 1000, arma::fill::randu);
+  // Generate labels.
+  arma::Row<size_t> labels(1000);
+  for (size_t i = 0; i < 500; ++i)
+    labels[i] = 0;
+  for (size_t i = 500; i < 1000; ++i)
+    labels[i] = 1;
+
+  // Create an empty tree.
+  HoeffdingTree<> ht;
+
+  // Just ensure that we can train without throwing an exception.
+  REQUIRE_NOTHROW(ht.Train(data, labels));
+
+  // Now, create a categorical dataset and retrain.
+  arma::mat data2 = arma::mat(4, 3000);
+  arma::Row<size_t> labels2(3000);
+  data::DatasetInfo info(4); // All features are numeric, except the fourth.
+  info.MapString<double>("0", 3);
+  for (size_t i = 0; i < 3000; i += 3)
+  {
+    data2(0, i) = mlpack::math::Random();
+    data2(1, i) = mlpack::math::Random();
+    data2(2, i) = mlpack::math::Random();
+    data2(3, i) = 0.0;
+    labels2[i] = 0;
+
+    data2(0, i + 1) = mlpack::math::Random();
+    data2(1, i + 1) = mlpack::math::Random() - 1.0;
+    data2(2, i + 1) = mlpack::math::Random() + 0.5;
+    data2(3, i + 1) = 0.0;
+    labels2[i + 1] = 2;
+
+    data2(0, i + 2) = mlpack::math::Random();
+    data2(1, i + 2) = mlpack::math::Random() + 1.0;
+    data2(2, i + 2) = mlpack::math::Random() + 0.8;
+    data2(3, i + 2) = 0.0;
+    labels2[i + 2] = 1;
+  }
+
+  // Ensure we can train without throwing an exception.
+  REQUIRE_NOTHROW(ht.Train(data2, info, labels2));
+
+  // Train while specifying the number of classes.
+  REQUIRE_NOTHROW(ht.Train(data, labels, false, true, 2));
+  REQUIRE_NOTHROW(ht.Train(data2, info, labels2, false, 3));
 }